pyproject.toml
[project]
name = "agentic-planning-eval"
version = "0.1.0"
description = "A framework for evaluating agentic planners"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mlflow>=2.0.0",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "dspy-ai",
    "a2a",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.9.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["experiments", "green_agent", "purple_agent", "helpful"]

# ------------------------------------------------------------------
# Ruff Configuration
# ------------------------------------------------------------------
[tool.ruff]
target-version = "py310"
exclude = ["purple_agent"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
